+++ backend/src/main/java/com/hjjang/backend/domain/chat/domain/entity/ChatRoom.java
@@ -1,5 +1,6 @@
-package com.hjjang.backend.domain.chat.domain;
+package com.hjjang.backend.domain.chat.domain.entity;
 
+import com.hjjang.backend.domain.user.entity.User;
 import com.hjjang.backend.global.domain.BaseTimeEntity;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -8,17 +9,22 @@ import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
 
+import static javax.persistence.FetchType.LAZY;
+
 @Getter
 @Entity
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Table(name = "chat_room")
 public class ChatRoom extends BaseTimeEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(name = "")
-    private String name;
+    @ManyToOne(fetch = LAZY)
+    @JoinColumn(name = "user_id")
+    private User seller;
+
 
 }
